Fix unit test for PEAR deps under composer v1
Initially the test was failing because some of the indirect deps in
`composer.lock` enforced a PHP constraint < `8.0.0`. I was able to
workaround that by a simple `composer1 update`, which pruned the
lockfile heavily.

Then I noticed the test was failing because when it tried to extract the
version from the return value of the native helper that was failing.
Which I eventually tracked down to a difference in behavior of the
`getPrettyVersion` depending on the PHP version:
 - PHP 7 returns: "2.4.1"
 - PHP 8 returns: "2.4.1@stable"

Fixing this test may be as simple as tweaking the anchoring regex. Or we
could drop support for PEAR, but I'm a little hesitant to do that until
we drop support for `composer` `1`, as that's the easy time to drop it.